+++ changelog.md
@@ -1,2 +1,3 @@
 - Fix friendly fire not doing anything in parties
 - Fix party settings command setting incorrect values
+- Remove Admin claiming commands as they are now part of Cadmus